refactor(collection): 将userId和problemId类型从uint64改为int64
为了增强代码的灵活性和兼容性，将collection模块中的userId和problemId类型从uint64改为int64，并在插入数据库时进行类型转换。

+++ internal/domain/collection/collection.go
@@ -111,7 +111,7 @@ func (c *Collection) Delete() error {
 	return &errors.NoError
 }
 
-func (c *Collection) UpdateUser(userIds []uint64) error {
+func (c *Collection) UpdateUser(userIds []int64) error {
 	var err error
 	options := c.toOption()
 	_, err = dao.CollectionStore.SelectOne(options)
@@ -126,7 +126,7 @@ func (c *Collection) UpdateUser(userIds []uint64) error {
 	for _, id := range userIds {
 		_, err = dao.CollectionUserStore.Insert(entity.CollectionUser{
 			CollectionId: c.Id,
-			UserId:       id,
+			UserId:       uint64(id),
 		})
 		if err != nil {
 			errs = append(errs, err)
@@ -138,7 +138,7 @@ func (c *Collection) UpdateUser(userIds []uint64) error {
 	return nil
 }
 
-func (c *Collection) UpdateProblem(problemIds []uint64) error {
+func (c *Collection) UpdateProblem(problemIds []int64) error {
 	var err error
 	options := c.toOption()
 	_, err = dao.CollectionStore.SelectOne(options)
@@ -154,7 +154,7 @@ func (c *Collection) UpdateProblem(problemIds []uint64) error {
 	for _, id := range problemIds {
 		_, err = dao.CollectionProblemStore.Insert(entity.CollectionProblem{
 			CollectionId: c.Id,
-			ProblemId:    id,
+			ProblemId:    uint64(id),
 			Serial:       serial,
 		})
 		if err != nil {
